requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
psycopg2==2.9.5
pycparser==2.21
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
